requirements.txt
langchain>=0.1.16
langchain-core>=0.3.15
langchain-google-genai>=2.0.0
langchain-community>=0.2.0
google-generativeai>=0.7.0
chromadb==0.4.15
pypdf2==3.0.1
python-docx==1.1.0
streamlit==1.28.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
nltk==3.8.1